deprecate hand and card class
